api-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "company-api-mcp"
version = "0.1.0"
description = "Searchable MCP server for Company API docs — context-efficient alternative to raw OpenAPI spec"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
company-api-mcp = "company_api_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/company_api_mcp"]
